# smart-ground-hub/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SmartGround API
  version: 2.0.0
  description: >
    REST API for IoT device management with MQTT-based device discovery and
    command signaling. Includes endpoints to manage device types, register
    devices, and trigger config pushes to SmartBox firmware via MQTT.

    ## Device Management

    All device operations use the flat `/api/devices` endpoint. Devices are
    always registered on a SmartBox (network access point) but are managed
    as a flat list with optional filtering by SmartBox or Range.
    Use **POST /api/devices** to register and immediately trigger a config push
    to the SmartBox firmware via MQTT.

    ## Breaking changes from v2.x

    - `POST /api/smart-boxes/{id}/devices` removed — use `POST /api/devices`
    - `GET /api/smart-boxes/{id}/devices` removed — use `GET /api/devices?smartBoxId=X`
    - `POST/DELETE /api/ranges/{rangeId}/devices/{deviceId}` removed — use `POST/DELETE /api/devices/{id}/range`
    - `SmartBoxResponse.deviceCount` removed
servers:
  - url: http://localhost:8080
    description: JVM App
  - url: http://localhost:8081
    description: Native App
security:
  - bearerAuth: []
paths:
...
  /api/ranges:
    get:
      summary: List all ranges
      operationId: listRanges
      tags: [Range]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Zero-based page number
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Number of items per page
      responses:
        '200':
          description: A list of ranges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangePageResponse'

    post:
      summary: Create a new range
      operationId: createRange
      tags: [Range]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRangeRequest'
      responses:
        '201':
          description: Range created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangeSummaryResponse'
        '400':
          description: Bad request
  /api/ranges/{id}:
    get:
      summary: Get range by ID
      operationId: getRange
      tags: [Range]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Range found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangeDetailResponse'
        '404':
          description: Range not found

    put:
      summary: Update a range
      operationId: updateRange
      tags: [Range]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRangeRequest'
      responses:
        '200':
          description: Range updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangeSummaryResponse'
        '404':
          description: Range not found

    delete:
      summary: Delete a range
      operationId: deleteRange
      tags: [Range]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Range deleted
        '404':
          description: Range not found
  /api/ranges/{id}/locked:
    patch:
      summary: Lock or unlock a range
      description: >
        When a range is locked, device assignment (POST/DELETE /api/ranges/{id}/devices/{deviceId})
        is rejected with 409. Commands to devices within the range are still forwarded.
        Locking is a safety mechanism to prevent accidental reconfiguration during an active session.
      operationId: setRangeLocked
      tags: [Range]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLockedRequest'
      responses:
        '200':
          description: Range lock state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangeSummaryResponse'
        '404':
          description: Range not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Events (SSE)
  # ─────────────────────────────────────────────────────────────────────────────
...
  /api/smart-boxes:
    get:
      summary: List all smart boxes
      operationId: listSmartBoxes
      tags: [SmartBox]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Zero-based page number
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Number of items per page
      responses:
        '200':
          description: A list of smart boxes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartBoxPageResponse'
  /api/smart-boxes/{id}:
    get:
      summary: Get smart box by ID
      operationId: getSmartBox
      tags: [SmartBox]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Smart box found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartBoxResponse'
        '404':
          description: Smart box not found
  /api/smart-boxes/{id}/alias:
    patch:
      summary: Set alias for a smart box
      operationId: setSmartBoxAlias
      tags: [SmartBox]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAliasRequest'
      responses:
        '200':
          description: Alias set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartBoxResponse'
        '404':
          description: Smart box not found
  /api/smart-boxes/{id}/push-config:
    post:
      summary: Manually trigger a config push to a smart box
      description: >
        Forces the backend to re-publish the full device configuration to the
        SmartBox via MQTT (topic: smartboxes/{mac}/config). Useful after
        changes to device types or when a box reconnects without triggering
        auto-push via discovery.
      operationId: pushSmartBoxConfig
      tags: [SmartBox]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Config push accepted (MQTT publish dispatched)
        '404':
          description: Smart box not found
components:
  schemas:
...
    DeviceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        alias:
          type: string
        smartBoxId:
          type: string
          format: uuid
          description: ID of the SmartBox this device is connected to
        groupId:
          type: string
          format: uuid
          description: Device type group
        groupName:
          type: string
          description: Name of the device type group
        deviceTypeId:
          type: string
          format: uuid
        deviceType:
          type: string
        command:
          type: string
          description: Firmware command (e.g. FIRE, ON)
        boxType:
          type: string
        blocked:
          type: boolean
        healthy:
          type: boolean
        signalDurationMs:
          type: integer
          description: How long the signal is held active
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Optional delay before signal fires
        pinConfig:
          $ref: '#/components/schemas/PinConfig'
        rangeId:
          type: string
          format: uuid
          nullable: true
...
    RangeSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        locked:
          type: boolean
          description: >
            true when the range is locked. Locked ranges reject device assignment changes.
            Use PATCH /api/ranges/{id}/locked to change this state.
        deviceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    RangeDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        locked:
          type: boolean
          description: >
            true when the range is locked. Locked ranges reject device assignment changes.
            Use PATCH /api/ranges/{id}/locked to change this state.
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DeviceResponse'
        createdAt:
          type: string
          format: date-time
    CreateRangeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    UpdateRangeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    SetLockedRequest:
      type: object
      required: [locked]
      properties:
        locked:
          type: boolean

    # ───── Events (SSE) ────────────────────────────────────────────────────────
...
    SmartBoxStatus:
      type: string
      enum: [ONLINE, OFFLINE, BLOCKED]
      description: SmartBox operational status

    # ───── Pin Configuration ───────────────────────────────────────────────────
    PinConfig:
      type: object
      required: [pin, mode]
      properties:
        pin:
          type: integer
          minimum: 0
          maximum: 40
          description: GPIO pin number on the Raspberry Pi Pico 2W (valid range 0–29 for user GPIOs; 0–40 reserved)
        mode:
          type: string
          enum: [OUTPUT, INPUT, INPUT_PULLUP, INPUT_PULLDOWN, PWM]
          description: GPIO mode as understood by the firmware
        pwmFrequencyHz:
          type: integer
          minimum: 1
          description: Required when mode is PWM. Sets the PWM carrier frequency.
        invertLogic:
          type: boolean
          default: false
          description: When true, the firmware inverts the signal (active-low wiring)

    # ───── Pagination ──────────────────────────────────────────────────────────
    PageMeta:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
    RangePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RangeSummaryResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'
    SmartBoxPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SmartBoxResponse'
        meta:
          $ref: '#/components/schemas/PageMeta'

    # ───── SmartBox ───────────────────────────────────────────────────────────
    SmartBoxResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        macAddress:
          type: string
        alias:
          type: string
        status:
          $ref: '#/components/schemas/SmartBoxStatus'
        firmwareVersion:
          type: string
        configSynced:
          type: boolean
          description: >
            true once the SmartBox has ACK'd the latest config push via
            smartboxes/{mac}/config/ack
...
    SetAliasRequest:
      type: object
      required: [alias]
      properties:
        alias:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT